+++ src/content/panorama/scripts/custom_game/loading_screen/loading_screen_tips.js
@@ -112,6 +112,9 @@ var tipsGeneral = [{
     },{
         img: 'file://{images}/custom_game/hints/hint_GoldBorder.png',
         txt: '#hintGoldBorder'
+    },{
+        img: 'file://{images}/spellicons/techies_suicide.png',
+        txt: '#hintAntiKamikaze'
     },
 ];